The configuration file resides at `src/main/azure/servicebus-emulator/config.json`.
The file was relocated to better support live reload.
Watching a resource for changes is brittle because it is updated
indirectly through changes in the source tree.
Propagation of such changes is not always reliable.

`src/test` was not chosen because QuarkusDevModeTests supports copying
files from `src/test/azure` to a temporary `src/main/azure` directory
used in the tests.